ONE SMALL STEP FOR MAN, ONE GIANT LEAP FOR MANKIND - Ladies and gentlemen, we have movement - Fixed checkEligibility so it actually lets pieces move

+++ src/Chess/Selector.java
@@ -34,9 +34,9 @@ public class Selector extends Rectangle{
 
 
     public double getArrayX() {
-        return x / 32;
+        return (int)(x / 32);
     }
     public double getArrayY() {
-        return y / 32;
+        return (int)(y / 32);
     }
 }
